% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdr-capstone-code.R
\name{clean_eq_data}
\alias{clean_eq_data}
\title{Convert earthquake data from raw to tidy}
\usage{
clean_eq_data(x)
}
\arguments{
\item{x}{is a valid earthquake dataset imported using read_tsv()}
}
\value{
Returns a fully tidied Tibble
}
\description{
This function applies all the earthquake data cleaning operations at once. It consolidates
dates, numericizes longitude and latitude, cleans the location names, and reduces the number
of columns.
}
