% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdr-capstone-code.R
\name{consolidate_date_column}
\alias{consolidate_date_column}
\title{Consolidate the date column}
\usage{
consolidate_date_column(x)
}
\arguments{
\item{x}{must be a valid earthquake dataset read in in the Tibble format}
}
\value{
The function returns a tibble with a new "date" column
}
\description{
This function takes the multiple date columns (day, year month, etc.) of the raw data
and converts it into a single clean lubridate::ymd() format. It fills in dates for earthquakes
that do not have precise dates available. It also readjusts BCE dates to work with Lubridate.
}
\details{
Errors are likely caused by inputting data that is not valid earthquake data in Tibble form.
}
