% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdr-capstone-code.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean up earthquake location column}
\usage{
eq_location_clean(x)
}
\arguments{
\item{x}{must be a valid earthquake dataset read in in the Tibble format}
}
\value{
Returns a tibble with new country and location_name columns
}
\description{
Applies several string manipulation functions to clean up the location_name column.
It first removes the country name from the beginning of the string. Then it converts to
title case. Then it removes white space.
}
