% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdr-capstone-code.R
\name{eq_map}
\alias{eq_map}
\title{Plots filtered earthquake data on a map and labels it with a selected annotation.}
\usage{
eq_map(data, annot_col = "popup_text")
}
\arguments{
\item{data}{is a valid earthquake tibble object}

\item{annot_col}{is a valid column in the tibble}
}
\value{
Returns a map with earthquake data plotted on it
}
\description{
Requires a valid earthquake tibble object as an input. Errors may be caused if the
specified annotation column does not exist or is misspelled. The default column is
a predefined popup_text column, which includes a number of variables.
}
