% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdr-capstone-code.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Plots a timeline of earthquakes}
\usage{
geom_timeline(df, mindate, maxdate, countries)
}
\arguments{
\item{df}{is a valid tidy earthquake tibble}

\item{mindate}{is the minimum date in "YYYY-MM-DD" format}

\item{maxdate}{is the maximum date in "YYYY-MM-DD" format}

\item{countries}{is a vector list of countries}
}
\value{
A timeilne plot of earthquakes
}
\description{
This function creates a plot of earthquakes by date. The user can specify a minimum
and maximum date and select the countries by which to filter.
}
\details{
Errors are likely due to an invalid tibble or incorrectly formatted inputs for the
other paremeters. Countries parameter only filters for countries matched in the tibble.
}
