% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdr-capstone-code.R
\name{numeric_lonlat}
\alias{numeric_lonlat}
\title{Function to ensure latitude and longitude are numeric class}
\usage{
numeric_lonlat(x)
}
\arguments{
\item{x}{must be a valid earthquake data tibble}
}
\value{
Same tibble with lon and lat in numeric format
}
\description{
Internal function that just ensures lon and lat are numeric.
}
