% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdr-capstone-code.R
\name{reduce_columns}
\alias{reduce_columns}
\title{Function to take out unnecessary columns -- many are redundant or riddles with NAs}
\usage{
reduce_columns(x)
}
\arguments{
\item{x}{must be a valid earthquake dataset read in in the Tibble format}
}
\value{
A tibble with a few select columns
}
\description{
Internal function to minimize number of columns in the earthquake tibble
}
